ApiConnection: On bad status code, log response
If we get an unexpected response code, log the full response body in
addition to attempting to parse it as an Api*Error. Sometimes errors are
returned in fields not included in Api*Error, causing Gson to discard
them. In such rare cases, being able to see the full JSON error response
in the logs should help quite a bit.